Conflicts:
	Sources/Internal/Render/Highlevel/RenderPass.cpp
	Sources/Internal/Render/Highlevel/Vegetation/VegetationGeometry.cpp
	Sources/Internal/Render/RHI/GLES2/rhi_GLES2.cpp

Former-commit-id: 282ef876b8f0bacca24bf45d82a3281559067893

+++ Tools/ResourceEditor/Classes/Qt/Scene/System/OwnersSignatureSystem.cpp
@@ -59,7 +59,7 @@ void OwnersSignatureSystem::ProcessCommand(const Command2 *command, bool redo)
 
 bool OwnersSignatureSystem::IsCommandIdValid(int _id)
 {
-	for (int i = 0; i < COUNT_OF(validIDs); ++i)
+	for (size_t i = 0; i < COUNT_OF(validIDs); ++i)
 	{
 		if(validIDs[i] == _id)
 		{
